refactor: internalController, externalController 패키지 분리
