+++ flight-service/src/main/java/com/oneul_tanda/flight_service/presentation/controller/external/AirlineController.java
@@ -1,4 +1,4 @@
-package com.oneul_tanda.flight_service.presentation.controller;
+package com.oneul_tanda.flight_service.presentation.controller.external;
 
 import com.oneul_tanda.flight_service.application.service.airline.AirlineService;
 import com.oneul_tanda.flight_service.presentation.dtos.airline.AirlineResponse;
