+++ flight-service/src/main/java/com/oneul_tanda/flight_service/presentation/controller/internal/InternalController.java
@@ -1,4 +1,4 @@
-package com.oneul_tanda.flight_service.presentation.controller;
+package com.oneul_tanda.flight_service.presentation.controller.internal;
 
 import com.oneul_tanda.flight_service.application.service.flight.FlightService;
 import java.util.UUID;
